restructuring the byte array manipulations to its own module
